Huge update
I would have broken this apart, but everything felt too interconnected
to break it apart into smaller commits. This adds some new stylings
with sass to make the site pop a little bit more, as well as actually
making use of angular with some simple routes. Load order of the
vendor file was changed up so that jQuery comes first, this way
Angular will make use of the full version of jQuery rather than its
own lighter version of it.