Major update: beta tester login
A beta tester login is implemented at a very
rudamentary level, providing a testing of a scoped
directive for printing out a list of units. Files
have been restructured for a single directives file
rather than breaking out directives into their own
tiny individual files. The directives jade files
themselves are now in the jade directory under
the directives sub-directory for increased clarity.
A nav bar is added to the front page along with as
always tons of refactoring.