add setup scripting and documentation
Made an acutal setup script, along with adding some
documentation about contributing to the project.

This commit also includes the beginning of the inclusion
of language configuration files. These files probably
won't see real use for some time. But the longer they get
to spend up is more time for them to be corrected or
improved. I'm unfortunately mono-lingual, so I just did
my best to translate but I'm sure there are mistranslations
or terms that would feel awkward to native speakers.