Improve selection highlighting in hangar
Meks are now highlighted with a bright green border around them.
It's just the default limegreen color. It would probably be
better to come up with a color that's more fitting with the sites
style, but it works for now. The display of the currently selected
mek is set to only display when a mek is selected (this is done to
avoid when the hanger is first visited, and angular's double curly
syntax shows up in the case of a null). New functions have been
defined in the hangarController to faciliate this.